Refactor AI Scenario handling: update schema to use stringID and introduce HackstackScenarioIDNode for improved scenario management in CampaignEditorView and SettingsTabView.
